services/Ecommerce-Core/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
psycopg2-binary
